+++ codegen.yml
@@ -1,0 +1,7 @@
+schema: http://backendonline.codebootcamp.co.kr/graphql
+generates:
+    ./src/commons/types/generated/types.ts:
+        plugins:
+            - typescript
+        config:
+            typesPrefix: I
