+++ package.json
@@ -6,7 +6,8 @@
         "dev": "next dev",
         "build": "next build",
         "start": "next start",
-        "lint": "next lint"
+        "lint": "next lint",
+        "generate": "graphql-codegen"
     },
     "dependencies": {
         "@apollo/client": "^3.7.3",
@@ -23,8 +24,10 @@
         "react-hook-form": "^7.41.5"
     },
     "devDependencies": {
-        "@types/react": "17.0.2",
+        "@graphql-codegen/cli": "^3.2.1",
+        "@graphql-codegen/typescript": "^3.0.1",
         "@types/node": "17.0.2",
+        "@types/react": "17.0.2",
         "eslint": "^8.34.0",
         "eslint-config-next": "^13.1.6",
         "typescript": "^4.9.5"
