+++ pages/boards/[boardId]/index.tsx
