+++ src/components/units/board/detail/BoardDetail.emotions.ts
