+++ src/components/units/board/detail/BoardDetail.typescript.ts
@@ -1,0 +1,8 @@
+import { IQuery } from "../../../../commons/types/generated/types";
+
+export interface IBoardDetailUI {
+    data?: Pick<IQuery, "fetchBoard">;
+    onDelete: () => void;
+    onClickMoveToBoardList: () => void;
+    onClickMoveToBoardEdit: () => void;
+}
