+++ src/components/units/board/list/BoardList.typescript.ts
@@ -1,0 +1,7 @@
+import { IQuery } from "../../../../commons/types/generated/types";
+
+export interface IBoardListUI {
+    onClickMoveToBoardNew: () => void;
+    onClickMoveToBoardDetail: (id: string) => void;
+    data?: Pick<IQuery, "fetchBoards">;
+}
